+++ packages/utils/src/index.ts
@@ -2,13 +2,19 @@ export { default as filterDataAttributes } from './filter-data-attributes';
 export { default as filterAriaAttributes } from './filter-aria-attributes';
 export { default as filterInvalidAttributes } from './filter-invalid-attributes';
 export { default as isNumberish } from './is-numberish';
-export { default as warnDeprecatedComponent } from './warn-deprecated-component';
-export { default as warnDeprecatedProp } from './warn-deprecated-prop';
+export {
+  warnDeprecatedComponent,
+  useWarnDeprecatedComponent,
+} from './warn-deprecated-component';
+export {
+  warnDeprecatedProp,
+  useWarnDeprecatedProp,
+} from './warn-deprecated-prop';
 export { default as createSequentialId } from './create-sequential-id';
 export { default as getFieldId } from './get-field-id';
 export { default as parseTime } from './parse-time';
 export { canUseDOM } from './helpers/exenv';
 export { default as SafeHTMLElement } from './helpers/safeHTMLElement';
-export { default as warning } from './warning';
+export { warning, useWarning } from './warning';
 
 export { default as version } from './version';