+++ packages/components/collapsible-motion/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @commercetools-uikit/collapsible-motion
 
+## 15.12.0
+
+### Patch Changes
+
+- [#2419](https://github.com/commercetools/ui-kit/pull/2419) [`3ed7853ce`](https://github.com/commercetools/ui-kit/commit/3ed7853ce5a59c20c347450d1f0190e996d43f9f) Thanks [@emmenko](https://github.com/emmenko)! - Update dependencies
+
+- Updated dependencies [[`3ed7853ce`](https://github.com/commercetools/ui-kit/commit/3ed7853ce5a59c20c347450d1f0190e996d43f9f), [`6e0ade9bb`](https://github.com/commercetools/ui-kit/commit/6e0ade9bb909e09bb40e7804b88e06f2f8f7e6d0)]:
+  - @commercetools-uikit/hooks@15.12.0
+  - @commercetools-uikit/utils@15.12.0
+
 ## 15.11.2
 
 ### Patch Changes